Move certificate handler
